Refactor Update screen tests
- testslater_btn_force_update_test() moved to UpdateViewTest.kt, as it does not need createAndroidComposeRule()
