+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/update/view/UpdateViewTest.kt
@@ -6,6 +6,7 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.onRoot
 import androidx.compose.ui.test.performClick
+import androidx.test.espresso.Espresso
 import androidx.test.filters.MediumTest
 import co.electriccoin.zcash.test.UiTestPrerequisites
 import co.electriccoin.zcash.ui.R
@@ -89,6 +90,28 @@ class UpdateViewTest : UiTestPrerequisites() {
         ).also {
             it.assertExists()
         }
+    }
+
+    @Test
+    @MediumTest
+    fun later_btn_force_update_test() {
+        val updateInfo = UpdateInfoFixture.new(
+            priority = AppUpdateChecker.Priority.HIGH,
+            force = true,
+            appUpdateInfo = null,
+            state = UpdateState.Prepared,
+        )
+        val testSetup = newTestSetup(updateInfo)
+
+        assertEquals(0, testSetup.getOnLaterCount())
+
+        composeTestRule.clickLater()
+
+        assertEquals(0, testSetup.getOnLaterCount())
+
+        Espresso.pressBack()
+
+        assertEquals(0, testSetup.getOnLaterCount())
     }
 
     @Test
